Attempt to make create no_std compatible
This is an initial attempt to make the crate no_std compatible.

Most of the features of the crate wouln't work, as neither `url` nor
`regex` are no_std compat yet.

It is more likely that `regex` becomes `no_std` compat than `url`, and
if that happens, we could make a new version of the parser without
adding the url attribute on the struct.